# api/public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exam Management API",
        "description": "API for managing exams, students, and attempt tracking",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/student/exams/{examId}": {
            "get": {
                "tags": [
                    "Student Exams"
                ],
                "summary": "Get student dashboard for an exam",
                "description": "Returns exam details, attempt status, and whether the student can start a new attempt",
                "operationId": "6605d67a6505d75bb1701af214a1ffe0",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam dashboard",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string"
                                        },
                                        "maxAttempts": {
                                            "type": "integer"
                                        },
                                        "attemptsRemaining": {
                                            "type": "integer"
                                        },
                                        "cooldownMinutes": {
                                            "type": "integer"
                                        },
                                        "cooldownUntil": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "canStart": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/exams/{examId}/start": {
            "post": {
                "tags": [
                    "Student Exams"
                ],
                "summary": "Start a new exam attempt",
                "operationId": "a2b8ab25e40bcaf225a899da03e299ac",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Attempt started",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "attemptId": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Exam not found"
                    },
                    "409": {
                        "description": "Cannot start attempt (no attempts left, cooldown active, or attempt in progress)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/attempts/{attemptId}/submit": {
            "post": {
                "tags": [
                    "Student Exams"
                ],
                "summary": "Submit an exam attempt",
                "operationId": "6b0a96b64276a15a48eb14f635cf39d5",
                "parameters": [
                    {
                        "name": "attemptId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Attempt submitted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Attempt not found"
                    },
                    "409": {
                        "description": "Invalid attempt state"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/attempts": {
            "get": {
                "tags": [
                    "Student Exams"
                ],
                "summary": "Get student attempt history",
                "operationId": "b9358c82e7e1a454322cbe88e604a1c4",
                "responses": {
                    "200": {
                        "description": "List of student attempts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "examId": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "examTitle": {
                                                "type": "string"
                                            },
                                            "attemptNumber": {
                                                "type": "integer"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "IN_PROGRESS",
                                                    "COMPLETED"
                                                ]
                                            },
                                            "startedAt": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "endedAt": {
                                                "type": "string",
                                                "format": "date-time",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/exams/{examId}/current-attempt": {
            "get": {
                "tags": [
                    "Student Exams"
                ],
                "summary": "Get current in-progress attempt for an exam",
                "operationId": "868dc07fade518e903ee2bd3fb222399",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Current attempt or null",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "currentAttempt": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "attemptNumber": {
                                                    "type": "integer"
                                                },
                                                "startedAt": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/admin/exams/{examId}/attempts": {
            "get": {
                "tags": [
                    "Admin Exams"
                ],
                "summary": "Get attempt history for an exam",
                "operationId": "58e82c6753a02cddda0fbe3cb23dbc9e",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of attempts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "studentId": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "studentName": {
                                                "type": "string"
                                            },
                                            "studentEmail": {
                                                "type": "string"
                                            },
                                            "attemptNumber": {
                                                "type": "integer"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "IN_PROGRESS",
                                                    "COMPLETED"
                                                ]
                                            },
                                            "startedAt": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "endedAt": {
                                                "type": "string",
                                                "format": "date-time",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using Bearer scheme",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}